Remove old parser
