+++ src/Language/Mimsa/Parser.hs
@@ -1,8 +1,6 @@
 module Language.Mimsa.Parser
   ( module Language.Mimsa.Parser.Language,
-    module Language.Mimsa.Parser.Parser,
   )
 where
 
 import Language.Mimsa.Parser.Language
-import Language.Mimsa.Parser.Parser
