+++ src/Language/Mimsa/Parser/Language.hs
@@ -2,16 +2,22 @@
 
 module Language.Mimsa.Parser.Language
   ( parseExpr,
-    parseExpr',
+    parseExprAndFormatError,
+    parseAndFormat,
     expressionParser,
+    varParser,
     nameParser,
     tyConParser,
     typeDeclParser,
+    ParseErrorType,
+    Parser,
+    thenSpace,
   )
 where
 
-import Control.Applicative ((<|>), optional)
 import Control.Monad ((>=>))
+import Data.Bifunctor (first)
+import Data.Char as Char
 import Data.Functor (($>))
 import Data.List.NonEmpty (NonEmpty)
 import qualified Data.List.NonEmpty as NE
@@ -21,8 +27,7 @@ import Data.Set (Set)
 import qualified Data.Set as S
 import Data.Text (Text)
 import qualified Data.Text as T
-import Language.Mimsa.Parser.Parser (Parser)
-import qualified Language.Mimsa.Parser.Parser as P
+import Data.Void
 import Language.Mimsa.Types
   ( DataType (DataType),
     Expr (..),
@@ -34,59 +39,85 @@ import Language.Mimsa.Types
     safeMkName,
     safeMkTyCon,
   )
+import Text.Megaparsec
+import Text.Megaparsec.Char
+import qualified Text.Megaparsec.Char.Lexer as L
+
+type Parser = Parsec Void Text
+
+type ParseErrorType = ParseErrorBundle Text Void
 
 type ParserExpr ann = Expr Name ann
 
+sc :: Parser ()
+sc =
+  L.space
+    space1 -- (2)
+    (L.skipLineComment "//") -- (3)
+    (L.skipBlockComment "/*" "*/") -- (4)
+
+lexeme :: Parser a -> Parser a
+lexeme = L.lexeme sc
+
+thenSpace :: Parser a -> Parser a
+thenSpace parser = do
+  _ <- space
+  val <- parser
+  _ <- space1
+  pure val
+
 -- parse expr, using it all up
-parseExpr :: Monoid ann => Text -> Either Text (ParserExpr ann)
-parseExpr input = P.runParser (P.thenOptionalSpace expressionParser) input
-  >>= \(leftover, a) ->
-    if T.length leftover == 0
-      then Right a
-      else Left ("Leftover input: >>>" <> leftover <> "<<<")
+parseExpr :: Monoid ann => Text -> Either ParseErrorType (ParserExpr ann)
+parseExpr = parse (expressionParser <* eof) "file.mimsa"
 
-parseExpr' :: Monoid ann => Text -> Either Text (ParserExpr ann)
-parseExpr' input = snd <$> P.runParser expressionParser input
+parseExprAndFormatError :: Monoid ann => Text -> Either Text (ParserExpr ann)
+parseExprAndFormatError = parseAndFormat (expressionParser <* eof)
 
-failer :: Parser (ParserExpr ann)
-failer = P.parseFail (\input -> "Could not parse expression for >>>" <> input <> "<<<")
+parseAndFormat :: Parser a -> Text -> Either Text a
+parseAndFormat p = first (T.pack . errorBundlePretty) . parse p "repl"
 
 expressionParser :: Monoid ann => Parser (ParserExpr ann)
 expressionParser =
   let parsers =
-        literalParser
-          <|> complexParser
-          <|> varParser
-          <|> constructorParser
+        try literalParser
+          <|> try complexParser
+          <|> try varParser
+          <|> try constructorParser
    in orInBrackets parsers
-        <|> failer
+
+between2 :: Char -> Char -> Parser a -> Parser a
+between2 a b parser = do
+  _ <- char a
+  val <- parser
+  _ <- char b
+  pure val
 
 inBrackets :: Parser a -> Parser a
-inBrackets = P.between2 '(' ')'
+inBrackets = between2 '(' ')'
 
 orInBrackets :: Parser a -> Parser a
-orInBrackets parser = parser <|> inBrackets parser
+orInBrackets parser = try parser <|> try (inBrackets parser)
 
 literalParser :: Monoid ann => Parser (ParserExpr ann)
 literalParser =
-  boolParser
-    <|> unitParser
-    <|> intParser
-    <|> stringParser
+  try boolParser
+    <|> try unitParser
+    <|> try intParser
+    <|> try stringParser
 
 complexParser :: Monoid ann => Parser (ParserExpr ann)
 complexParser =
-  letPairParser
-    <|> recordParser
-    <|> letParser
-    <|> ifParser
-    <|> appParser
-    <|> pairParser
-    <|> recordAccessParser
-    <|> lambdaParser
-    <|> typeParser
-    <|> constructorAppParser
-    <|> caseMatchParser
+  try letPairParser
+    <|> try recordParser
+    <|> try letParser
+    <|> try ifParser
+    <|> try appParser
+    <|> try pairParser
+    <|> try recordAccessParser
+    <|> try lambdaParser
+    <|> try typeParser
+    <|> try constructorAppParser
+    <|> try caseMatchParser
 
 protectedNames :: Set Text
 protectedNames =
@@ -107,29 +138,44 @@ protectedNames =
 
 ----
 
+integer' :: Parser Int
+integer' = lexeme L.decimal
+
+integerParser :: Parser Int
+integerParser = L.signed sc integer'
+
+---
+
 boolParser :: Monoid ann => Parser (ParserExpr ann)
 boolParser = trueParser <|> falseParser
 
 trueParser :: Monoid ann => Parser (ParserExpr ann)
-trueParser = P.literal "True" $> MyLiteral mempty (MyBool True)
+trueParser = string "True" $> MyLiteral mempty (MyBool True)
 
 falseParser :: Monoid ann => Parser (ParserExpr ann)
-falseParser = P.literal "False" $> MyLiteral mempty (MyBool False)
+falseParser = string "False" $> MyLiteral mempty (MyBool False)
 
 -----
 
 unitParser :: Monoid ann => Parser (ParserExpr ann)
-unitParser = P.literal "Unit" $> MyLiteral mempty MyUnit
+unitParser = string "Unit" $> MyLiteral mempty MyUnit
 
 -----
 
 intParser :: Monoid ann => Parser (ParserExpr ann)
-intParser = MyLiteral mempty . MyInt <$> P.integer
+intParser = MyLiteral mempty . MyInt <$> integerParser
 
 -----
 
+stringLiteral :: Parser String
+stringLiteral = char '\"' *> manyTill L.charLiteral (char '\"')
+
 stringParser :: Monoid ann => Parser (ParserExpr ann)
-stringParser = MyLiteral mempty . MyString . StringType <$> P.between '"'
+stringParser =
+  MyLiteral mempty . MyString
+    . StringType
+    . T.pack
+    <$> stringLiteral
 
 -----
 
@@ -138,14 +184,27 @@ varParser = MyVar mempty <$> nameParser
 
 ---
 
+identifier :: Parser Text
+identifier = takeWhile1P (Just "variable name") Char.isAlphaNum
+
+maybePred :: (Show a) => Parser a -> (a -> Maybe b) -> Parser b
+maybePred parser predicate' = try $ do
+  a <- parser
+  case predicate' a of
+    Just b -> pure b
+    _ -> fail $ T.unpack $ "Predicate did not hold for " <> T.pack (show a)
+
 nameParser :: Parser Name
 nameParser =
-  P.maybePred
-    P.identifier
+  maybePred
+    identifier
     (inProtected >=> safeMkName)
 
 inProtected :: Text -> Maybe Text
-inProtected tx = if S.member tx protectedNames then Nothing else Just tx
+inProtected tx =
+  if S.member tx protectedNames
+    then Nothing
+    else Just tx
 
 ---
 
@@ -154,27 +213,27 @@ constructorParser = MyConstructor mempty <$> tyConParser
 
 tyConParser :: Parser TyCon
 tyConParser =
-  P.maybePred
-    P.identifier
+  maybePred
+    identifier
     (inProtected >=> safeMkTyCon)
 
 -----
 
 letParser :: Monoid ann => Parser (ParserExpr ann)
-letParser = letInParser <|> letNewlineParser
+letParser = try letInParser <|> letNewlineParser
 
 letNameIn :: Parser Name
 letNameIn = do
-  _ <- P.thenSpace (P.literal "let")
-  name <- P.thenSpace nameParser
-  _ <- P.thenSpace (P.literal "=")
+  _ <- thenSpace (string "let")
+  name <- thenSpace nameParser
+  _ <- thenSpace (string "=")
   pure name
 
 letInParser :: Monoid ann => Parser (ParserExpr ann)
 letInParser = do
   name <- letNameIn
-  expr <- P.thenSpace expressionParser
-  _ <- P.thenSpace (P.literal "in")
+  expr <- withOptionalSpace expressionParser
+  _ <- thenSpace (string "in")
   MyLet mempty name expr <$> expressionParser
 
 letNewlineParser :: Monoid ann => Parser (ParserExpr ann)
@@ -188,9 +247,9 @@ letNewlineParser = do
 
 spacedName :: Parser Name
 spacedName = do
-  _ <- P.space0
+  _ <- space
   name <- nameParser
-  _ <- P.space0
+  _ <- space
   pure name
 
 -----
@@ -202,23 +261,26 @@ letPairParser =
     <*> inParser
   where
     binder1 = do
-      _ <- P.thenSpace (P.literal "let")
-      _ <- P.literal "("
+      _ <- thenSpace (string "let")
+      _ <- string "("
       spacedName
     binder2 = do
-      _ <- P.literal ","
+      _ <- string ","
       name <- spacedName
-      _ <- P.thenSpace (P.literal ")")
+      _ <- thenSpace (string ")")
       pure name
 
 -----
 
 equalsParser :: Monoid ann => Parser (ParserExpr ann)
-equalsParser =
-  P.right (P.thenSpace (P.literal "=")) (P.thenSpace expressionParser)
+equalsParser = do
+  _ <- thenSpace (string "=")
+  thenSpace expressionParser
 
 inParser :: Monoid ann => Parser (ParserExpr ann)
-inParser = P.right (P.thenSpace (P.literal "in")) expressionParser
+inParser = do
+  _ <- thenSpace (string "in")
+  expressionParser
 
 -----
 
@@ -228,32 +290,36 @@ lambdaParser = MyLambda mempty <$> slashNameBinder <*> arrowExprBinder
 -- matches \varName
 slashNameBinder :: Parser Name
 slashNameBinder = do
-  _ <- P.literal "\\"
-  _ <- P.space0
-  P.thenSpace nameParser
+  _ <- string "\\"
+  _ <- space
+  thenSpace nameParser
 
 arrowExprBinder :: Monoid ann => Parser (ParserExpr ann)
-arrowExprBinder = P.right (P.thenSpace (P.literal "->")) expressionParser
+arrowExprBinder = do
+  _ <- thenSpace (string "->")
+  expressionParser
 
 -----
 
 appFunc :: Monoid ann => Parser (ParserExpr ann)
-appFunc = recordAccessParser <|> inBrackets lambdaParser <|> varParser
+appFunc =
+  try recordAccessParser <|> try (inBrackets lambdaParser)
+    <|> varParser
 
 appParser :: Monoid ann => Parser (ParserExpr ann)
 appParser = do
-  app <- appFunc
-  exprs <- P.oneOrMore (withOptionalSpace exprInBrackets)
-  pure (foldl (MyApp mempty) app exprs)
+  func <- appFunc
+  exprs <- some (withOptionalSpace exprInBrackets)
+  pure (foldl (MyApp mempty) func exprs)
 
 literalWithSpace :: Text -> Parser ()
-literalWithSpace tx = () <$ withOptionalSpace (P.literal tx)
+literalWithSpace tx = () <$ withOptionalSpace (string tx)
 
 withOptionalSpace :: Parser a -> Parser a
 withOptionalSpace p = do
-  _ <- P.space0
+  _ <- space
   a <- p
-  _ <- P.space0
+  _ <- space
   pure a
 
 exprInBrackets :: Monoid ann => Parser (ParserExpr ann)
@@ -261,27 +327,17 @@ exprInBrackets = do
   literalWithSpace "("
   expr <- expressionParser
   literalWithSpace ")"
-  _ <- P.space0
+  _ <- space
   pure expr
 
 -----
 
 recordParser :: Monoid ann => Parser (ParserExpr ann)
-recordParser = fullRecordParser <|> emptyRecordParser
-
-emptyRecordParser :: Monoid ann => Parser (ParserExpr ann)
-emptyRecordParser = do
+recordParser = do
   literalWithSpace "{"
+  args <- sepBy (withOptionalSpace recordItemParser) (literalWithSpace ",")
   literalWithSpace "}"
-  pure (MyRecord mempty mempty)
-
-fullRecordParser :: Monoid ann => Parser (ParserExpr ann)
-fullRecordParser = do
-  literalWithSpace "{"
-  args <- P.zeroOrMore (P.left recordItemParser (P.left (P.literal ",") P.space0))
-  last' <- recordItemParser
-  literalWithSpace "}"
-  pure (MyRecord mempty (M.fromList $ args <> [last']))
+  pure (MyRecord mempty (M.fromList args))
 
 recordItemParser :: Monoid ann => Parser (Name, ParserExpr ann)
 recordItemParser = do
@@ -294,20 +350,20 @@ recordItemParser = do
 
 recordAccessParser :: Monoid ann => Parser (ParserExpr ann)
 recordAccessParser =
-  recordAccessParser3
-    <|> recordAccessParser2
-    <|> recordAccessParser1
+  try recordAccessParser3
+    <|> try recordAccessParser2
+    <|> try recordAccessParser1
 
 recordAccessParser1 :: Monoid ann => Parser (ParserExpr ann)
 recordAccessParser1 = do
   expr <- varParser
   name <- dotName
-  _ <- P.space0
+  _ <- space
   pure (MyRecordAccess mempty expr name)
 
 dotName :: Parser Name
 dotName = do
-  _ <- P.literal "."
+  _ <- string "."
   nameParser
 
 recordAccessParser2 :: Monoid ann => Parser (ParserExpr ann)
@@ -315,19 +371,19 @@ recordAccessParser2 = do
   expr <- varParser
   name <- dotName
   name2 <- dotName
-  _ <- P.space0
+  _ <- space
   pure (MyRecordAccess mempty (MyRecordAccess mempty expr name) name2)
 
 recordAccessParser3 :: Monoid ann => Parser (ParserExpr ann)
 recordAccessParser3 = do
   expr <- varParser
-  _ <- P.literal "."
+  _ <- string "."
   name <- nameParser
-  _ <- P.literal "."
+  _ <- string "."
   name2 <- nameParser
-  _ <- P.literal "."
+  _ <- string "."
   name3 <- nameParser
-  _ <- P.space0
+  _ <- space
   pure (MyRecordAccess mempty (MyRecordAccess mempty (MyRecordAccess mempty expr name) name2) name3)
 
 -----
@@ -336,56 +392,47 @@ ifParser :: Monoid ann => Parser (ParserExpr ann)
 ifParser = MyIf mempty <$> predParser <*> thenParser <*> elseParser
 
 predParser :: Monoid ann => Parser (ParserExpr ann)
-predParser = P.right (P.thenSpace (P.literal "if")) expressionParser
+predParser = do
+  _ <- thenSpace (string "if")
+  expressionParser
 
 thenParser :: Monoid ann => Parser (ParserExpr ann)
-thenParser = P.right (P.thenSpace (P.literal "then")) expressionParser
+thenParser = do
+  _ <- thenSpace (string "then")
+  expressionParser
 
 elseParser :: Monoid ann => Parser (ParserExpr ann)
-elseParser = P.right (P.thenSpace (P.literal "else")) expressionParser
+elseParser = do
+  _ <- thenSpace (string "else")
+  expressionParser
 
 -----
 
 pairParser :: Monoid ann => Parser (ParserExpr ann)
 pairParser = do
-  _ <- P.literal "("
-  _ <- P.space0
+  _ <- string "("
+  _ <- space
   exprA <- expressionParser
-  _ <- P.space0
-  _ <- P.literal ","
-  _ <- P.space0
+  _ <- space
+  _ <- string ","
+  _ <- space
   exprB <- expressionParser
-  _ <- P.space0
-  _ <- P.literal ")"
-  _ <- P.space0
+  _ <- space
+  _ <- string ")"
+  _ <- space
   pure $ MyPair mempty exprA exprB
 
 -----
 
-typeDeclParser :: Parser DataType
-typeDeclParser = typeDeclParserWithCons <|> typeDeclParserEmpty
-
-typeDeclParserEmpty :: Parser DataType
-typeDeclParserEmpty = do
-  _ <- P.thenSpace (P.literal "type")
-  tyName <- tyConParser
-  pure (DataType tyName mempty mempty)
-
-typeDeclParserWithCons :: Parser DataType
-typeDeclParserWithCons = do
-  _ <- P.thenSpace (P.literal "type")
-  tyName <- P.thenSpace tyConParser
-  tyArgs <- P.zeroOrMore (P.left nameParser P.space1)
-  _ <- P.thenSpace (P.literal "=")
-  constructors <- manyTypeConstructors <|> oneTypeConstructor
-  pure $ DataType tyName tyArgs constructors
-
---------
-
 typeParser :: Monoid ann => Parser (ParserExpr ann)
 typeParser =
-  MyData mempty <$> (typeDeclParserWithCons <|> typeDeclParserEmpty)
-    <*> (inExpr <|> inNewLineExpr)
+  MyData mempty
+    <$> ( try typeDeclParserWithCons
+            <|> try typeDeclParserEmpty
+        )
+    <*> ( try inExpr
+            <|> try inNewLineExpr
+        )
 
 inNewLineExpr :: Monoid ann => Parser (ParserExpr ann)
 inNewLineExpr = do
@@ -394,37 +441,77 @@ inNewLineExpr = do
 
 inExpr :: Monoid ann => Parser (ParserExpr ann)
 inExpr = do
-  _ <- P.space0
-  _ <- P.thenSpace (P.literal "in")
+  _ <- space
+  _ <- thenSpace (string "in")
   expressionParser
+
+-----
+
+typeDeclParser :: Parser DataType
+typeDeclParser =
+  try typeDeclParserWithCons
+    <|> try typeDeclParserEmpty
+
+-- it's your "type Void in ..."
+typeDeclParserEmpty :: Parser DataType
+typeDeclParserEmpty = do
+  _ <- thenSpace (string "type")
+  tyName <- tyConParser
+  pure (DataType tyName mempty mempty)
+
+-- it's your more complex cases
+typeDeclParserWithCons :: Parser DataType
+typeDeclParserWithCons = do
+  _ <- thenSpace (string "type")
+  tyName <- thenSpace tyConParser
+  tyArgs <- try $ many (thenSpace nameParser)
+  _ <- thenSpace (string "=")
+  constructors <-
+    try manyTypeConstructors
+      <|> try oneTypeConstructor
+  pure $ DataType tyName tyArgs constructors
+
+--------
 
 ----
 
 manyTypeConstructors :: Parser (Map TyCon [TypeName])
 manyTypeConstructors = do
-  cons <- NE.toList <$> P.oneOrMore (P.left oneTypeConstructor (P.thenSpace (P.literal "|")))
-  lastCons <- oneTypeConstructor
-  pure (mconcat cons <> lastCons)
+  tyCons <-
+    sepBy
+      (withOptionalSpace oneTypeConstructor)
+      (literalWithSpace "|")
+  pure (mconcat tyCons)
 
 oneTypeConstructor :: Parser (Map TyCon [TypeName])
 oneTypeConstructor = do
   name <- tyConParser
-  args <- P.zeroOrMore (P.right P.space1 typeNameParser)
+  args <-
+    try
+      ( do
+          _ <- space1
+          sepBy (withOptionalSpace typeNameParser) space
+      )
+      <|> pure mempty
   pure (M.singleton name args)
 
 -----
 
 typeNameParser :: Parser TypeName
 typeNameParser =
-  emptyConsParser <|> varNameParser <|> inBrackets parameterisedConsParser <|> varNameParser
+  try emptyConsParser
+    <|> try varNameParser
+    <|> try (inBrackets parameterisedConsParser)
 
+-- Simple type like String
 emptyConsParser :: Parser TypeName
 emptyConsParser = ConsName <$> tyConParser <*> pure mempty
 
+--
 parameterisedConsParser :: Parser TypeName
 parameterisedConsParser = do
-  c <- tyConParser
-  params <- P.zeroOrMore (P.right P.space1 typeNameParser)
+  c <- thenSpace tyConParser
+  params <- try (sepBy (withOptionalSpace typeNameParser) space1) <|> pure mempty
   pure $ ConsName c params
 
 varNameParser :: Parser TypeName
@@ -435,22 +522,26 @@ varNameParser = VarName <$> nameParser
 constructorAppParser :: Monoid ann => Parser (ParserExpr ann)
 constructorAppParser = do
   cons <- tyConParser
-  exprs <- P.oneOrMore (P.right P.space1 (orInBrackets expressionParser))
+  exprs <-
+    sepBy
+      (withOptionalSpace (orInBrackets expressionParser))
+      space
   pure (foldl (MyConsApp mempty) (MyConstructor mempty cons) exprs)
 
 ----------
 caseExprOfParser :: Monoid ann => Parser (ParserExpr ann)
 caseExprOfParser = do
-  _ <- P.thenSpace (P.literal "case")
+  _ <- thenSpace (string "case")
   sumExpr <- expressionParser
-  _ <- P.thenSpace (P.literal "of")
+  _ <- thenSpace (string "of")
   pure sumExpr
 
 caseMatchParser :: Monoid ann => Parser (ParserExpr ann)
 caseMatchParser = do
   sumExpr <- caseExprOfParser
   matches <-
-    matchesParser <|> pure <$> matchParser
+    try matchesParser
+      <|> pure <$> matchParser
   catchAll <-
     optional (otherwiseParser (not . null $ matches))
   pure $ MyCaseMatch mempty sumExpr matches catchAll
@@ -458,16 +549,17 @@ caseMatchParser = do
 otherwiseParser :: Monoid ann => Bool -> Parser (ParserExpr ann)
 otherwiseParser needsBar = do
   if needsBar
-    then () <$ P.thenSpace (P.literal "|")
+    then () <$ thenSpace (string "|")
     else pure ()
-  _ <- P.thenSpace (P.literal "otherwise")
+  _ <- thenSpace (string "otherwise")
   expressionParser
 
 matchesParser :: Monoid ann => Parser (NonEmpty (TyCon, ParserExpr ann))
-matchesParser = do
-  cons <- P.zeroOrMore (P.left matchParser (P.thenSpace (P.literal "|")))
-  lastCons <- matchParser
-  pure $ NE.fromList (cons <> [lastCons])
+matchesParser =
+  NE.fromList
+    <$> sepBy
+      (withOptionalSpace matchParser)
+      (literalWithSpace "|")
 
 matchParser :: Monoid ann => Parser (TyCon, ParserExpr ann)
-matchParser = (,) <$> P.thenSpace tyConParser <*> expressionParser
+matchParser = (,) <$> thenSpace tyConParser <*> expressionParser
